requirements.txt
flask
flask-cors
pytorch==1.10
scikit-learn
numpy
altair
gym
pfrl
